[AUTO] Update dependencies
